a2-backend-BookmarkAPI/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
flask_cors==3.0.8
google-cloud-datastore==1.12.0
